+++ CHANGELOG.md
@@ -12,8 +12,8 @@
 
 - Use bract.core 0.6.2-beta1
 - [Todo] Move `bract.dev.ring` stuff here in `bract.ring.dev`
-- [Todo] Add support for [nginx-clojure-embedded](https://github.com/nginx-clojure/nginx-clojure/tree/master/nginx-clojure-embed)
-  - [Todo] bract.ring.server/start-nginx-clojure-embedded-server
+- Add support for [nginx-clojure-embedded](https://github.com/nginx-clojure/nginx-clojure/tree/master/nginx-clojure-embed)
+  - `bract.ring.server/start-nginx-clojure-embedded-server`
 
 
 ## 0.6.2-0.2.0-alpha3 / 2020-October-12